Binance Futures Order Bot/requirements.txt
python-binance
python-dotenv
tabulate
